changed iddok from long to integer

+++ src/main/java/org/eotpremnice/entity/id/KurirEntityId.java
@@ -27,5 +27,5 @@ public class KurirEntityId implements Serializable {
     private String tipDokumenta;
 
     @Column(name = "IDDOK")
-    private Long iddok;
+    private Integer iddok;
 }
