+++ src/main/java/org/eotpremnice/entity/id/MagacinEntityId.java
@@ -27,5 +27,5 @@ public class MagacinEntityId implements Serializable {
     private String tipDokumenta;
 
     @Column(name = "IDDOK")
-    private Long iddok;
+    private Integer iddok;
 }
