+++ src/main/java/org/eotpremnice/entity/id/OdredisteEntityId.java
@@ -27,5 +27,5 @@ public class OdredisteEntityId implements Serializable {
     private String tipDokumenta;
 
     @Column(name = "IDDOK")
-    private Long iddok;
+    private Integer iddok;
 }
