+++ src/main/java/org/eotpremnice/entity/id/VozacEntityId.java
@@ -27,5 +27,5 @@ public class VozacEntityId implements Serializable {
     private String tipDokumenta;
 
     @Column(name = "IDDOK")
-    private Long iddok;
+    private Integer iddok;
 }
